requirements.txt
absl-py
astral
bt_proximity
git+https://github.com/pybluez/pybluez@4d46ce14d9e888e3b0c65d5d3ba2a703f8e5e861#egg=pybluez
pyyaml
requests
